+++ main-server/src/core/common/index.js
@@ -156,6 +156,8 @@ export function getFunction(fnName) {
       return closeRequestInternalAsyncAfterBlockchain;
     case 'common.isRequestClosedOrTimedOut':
       return isRequestClosedOrTimedOut;
+    case 'common.timeoutRequestAfterBlockchain':
+      return timeoutRequestAfterBlockchain;
     case 'idp.requestChallengeAfterBlockchain':
       return idp.requestChallengeAfterBlockchain;
     case 'idp.createResponseAfterBlockchain':
@@ -353,17 +355,42 @@ export async function timeoutRequest(nodeId, requestId) {
 
     await tendermintNdid.timeoutRequest(
       { requestId, responseValidList },
-      nodeId
+      nodeId,
+      'common.timeoutRequestAfterBlockchain',
+      [
+        {
+          nodeId,
+          requestId,
+        },
+      ],
+      true
     );
   } catch (error) {
     logger.error({
-      message: 'Cannot set timed out',
+      message: 'Cannot set timeout request',
       requestId,
       error,
     });
     throw error;
   }
-  cacheDb.removeTimeoutScheduler(nodeId, requestId);
+}
+
+function timeoutRequestAfterBlockchain(
+  { error, chainDisabledRetryLater },
+  { nodeId, requestId }
+) {
+  if (chainDisabledRetryLater) return;
+  try {
+    if (error) throw error;
+    cacheDb.removeTimeoutScheduler(nodeId, requestId);
+  } catch (error) {
+    logger.error({
+      message: 'Timeout request after blockchain error',
+      tendermintResult: arguments[0],
+      additionalArgs: arguments[1],
+      error,
+    });
+  }
 }
 
 export function runTimeoutScheduler(nodeId, requestId, secondsToTimeout) {