+++ main-server/src/core/idp/create_response.js
@@ -360,9 +360,10 @@ export async function createResponse(
 }
 
 export async function createResponseAfterBlockchain(
-  { height, error },
+  { height, error, chainDisabledRetryLater },
   { nodeId, reference_id, callback_url, request_id, mode, privateProofObject }
 ) {
+  if (chainDisabledRetryLater) return;
   try {
     if (error) throw error;
 
@@ -464,7 +465,7 @@ async function requestChallenge({
 }
 
 export async function requestChallengeAfterBlockchain(
-  { height, error },
+  { height, error, chainDisabledRetryLater },
   {
     nodeId,
     reference_id,
@@ -476,6 +477,7 @@ export async function requestChallengeAfterBlockchain(
     rp_id,
   }
 ) {
+  if (chainDisabledRetryLater) return;
   try {
     if (error) throw error;
     //send message queue with public proof